requirements.txt
app-version==1.0.1
asgiref==3.8.1
Django==5.1.7
django-admin-interface==0.29.4
django-appconf==1.1.0
django-colorfield==0.12.0
django-jalali-date==1.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
jalali_core==1.0.0
jdatetime==5.2.0
mysqlclient==2.2.7
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.1
